Add admin client and finish cluster networking
The admin client communicates over tcp using blocking sockets and the
resp protocol, and talks to the admin port of v2r2. It provides config
get/set for the connected node as well as membership and cluster status.
By default the admin-cli runs as an interactive shell. It can run a
single command via the `-e` flag. This is useful for shell scripting.

After launching a development cluster with `make launch`, an admin
client connected to the first node (dev1) can be run via:
`cargo run --bin v2r2-admin -- 127.0.0.1:2001`. It's handy to wrap this
command using `rlwrap` so that you get typical readline support.

Admin status and membership are tracked as part of the global `State`
struct present in all handlers.

A 1 second tick was added to event loops and results in the periodic
firing of the `tick` callback in server handlers. Currently it is only
used in the cluster server and serves the following purposes:

 * Connect to peers that are members of the cluster, but not currently
   connected (auto-reconnect)
 * Remove connected peers that are no longer members of the cluster
 * Inform the membership system of connected/disconnected peers so that
   it can update the global state used by the admin server for
   reporting.
 * Send `Ping` messages to peers to detect liveness (heartbeating)