Add an API for VR along with a CLI client
The API is currently message pack wrapped in Resp bulk strings for
framing. It's expected that it will be transitioned to another framing
mechanism such as a 4 byte header.

The API implements only a few commands, and so far only binary
representations actually work. The implemented commands are:

 * Create - Create an element of a given type at some path
 * Get
 * Put (w/ CAS as an option)
 * Delete (w/ CAS as an option)
 * List - List keys under a given path

The API uses a mock VR layer that exists in it's own thread and receives
messages on a channel from the API handler. The mock VR layer doesn't
actually do anything but make upcalls to the backend data structure that
implements the API. Currently the backend uses a BTree to maintain
state, although it's expected that this will be replaced with some other
data structure that is easily snapshottable and able to be efficiently
persisted to disk. Each node maintains it's own separate btree, and
since the VR layer is just a stub, they are currently all independent
meaning that sending different operations to different nodes will result
in different responses.