Update tests to work with distill 0.2.0
Tests now deal with symbolic and dynamic operations. Tests also no
longer need to track history and record it for use by the debugger as
distill will do that. Tests also no longer need to track causal ids or
worry about directly dropping messages, as they can determine invalid
messages when converting from symbolic to dynamic messages in the update
function based on model state. This makes tests shorter, and simpler to
write.

The debugger code has also changed to no longer use 'Frames' and
'Actions' and now uses only dynamic ops. This allows it to be more
generic in the hopes that we can fully paramterize it and move it into
distill.