WIP: Start to refactor FSM bigtime
Move VrCtx into its own module. Start to put all specific mutation, and
message generating code in there as methods on VrCtx.

Reduce vr_fsm to the minimum amount of code required for state
transitions and global epoch/view checks. Try to ensure that all state
transitions utilizing the `next!` macro only occur in vr_fsm,
particularly in the state functions themselves. There are still a few
state transitions living in VrCtx. Those VrCtx methods need to be
modified to only mutate/return data and give the state functions the
information they need to decide which states to transition to.

Overall, the code isn't even close to compiling and this is very much a
WIP, but it's a good stab at getting organized and making the state
transition code much clearer to the reader of vr_fsm.

Lastly, It's hoped that many of the large number of methods moved into
vr_ctx.rs can be collapsed into more generic functionality.