Add checking of fsm constraints to test runs.
The start to a set of funfsm constraints to check invariants of each
vr_fsm was created in tests/utils/vr_fsm_constraints.rs. The constraints
are nowhere near complete. The only transition covered currently is
primary -> primary. The structure should make them easy to add to
however. Even with the small set of constraints created, testing has
already found a few violations and the fixes are included in this PR.

The scheduler was edited to ensure that invariants are checked before
and after every fsm callback during testing.