Add support for recovery during reconfiguration
Also change nonces to be u64. Note that this requires either storing nonces on disk or other nodes.
In the next commit, disk support will be provided for this purpose. An RFC will be required for
storing on other nodes. The mechanism used in the RFC should also fix the recovery problems
illustrated in this paper: https://drkp.net/papers/diskless-tr16.pdf