Implement view change optimizations in RFC 2
Additionally add some other protocol fixes/enhancements:
 1. Ensure PrepareOk message for last log entry gets sent after state transfer
 2. Rebroadcast prepare if primary has not received quorum
 3. Send PrepareOk for last log entry after view change and recovery

Some notes about the above enhancements:

 2. When there are outstanding uncommitted operations, and a primary timeout
    occurs resend the prepare instead of a commit message to ensure it
    eventually commits even if no more new client requests are received.

    Track the broadcast in `PrepareRequests` using the pid of the primary if
    it wasn't the one to receive the client request, such as after a view
    change. On commit filter on this pid so that a client reply is not send from
    the primary to itself.

    If a backup receives a Prepare message for the last operation in it's
    log it sends a PrepareOk so that the primary can commit in case prior
    PrepareOk messages were dropped.

 3. When becoming a backup, send a PrepareOk message for the last log entry
    to the primary. This is specified in the paper for view change, but not
    recovery. For recovery this allows the primary to update it's
    min_accepted map without having to wait for a new client request which
    may not come for quite a while. Note that this message can also be
    dropped, but as it is an optimization it is still safe.

Add a model of the VR replicas for QC tests

An abstract model of the state of a VR replica set has been created for use in quickcheck tests.
The state of all the replicas is compared against the model after each operation to ensure
correctness.

The primary idle timeout has been moved from the primary state into the context so that any changes
to its configuration can carry across state transitions.

Add a history member to the scheduler that allows us to track messages and operations throughout
tests and then only show the failing history. This is especially useful in QC tests when we only
want to display the history for the shrunk test.

Fix make-devrel to use the right binary to generate configuration.